+++ wenet/transformer/label_smoothing_loss.py
@@ -91,6 +91,6 @@ class LabelSmoothingLoss(nn.Module):
         total = len(target) - ignore.sum().item()
         target = target.masked_fill(ignore, 0)  # avoid -1 index
         true_dist.scatter_(1, target.unsqueeze(1), self.confidence)
-        kl = self.criterion(torch.log_softmax(x, dim=1), true_dist)
+        kl = self.criterion(torch.log_softmax(x, dim=1, dtype=torch.float32), true_dist)
         denom = total if self.normalize_length else batch_size
         return kl.masked_fill(ignore.unsqueeze(1), 0).sum() / denom
